Endret navnet på steget til virksomheter
