+++ package.json
@@ -20,7 +20,7 @@
     "lodash": "^4.17.11",
     "lodash.throttle": "^4.1.1",
     "melosys-ikoner-assets": "^1.0.1",
-    "melosys-kodeverk": "4.4.0",
+    "melosys-kodeverk": "4.6.2",
     "moment": "^2.24.0",
     "moment-duration-format": "^2.2.2",
     "nav-frontend-alertstriper": "^1.0.34",
