+++ src/soknad-komponenter/stegvelger/stegMotor/typer.js
@@ -25,14 +25,14 @@ export const STEG = {
   YRKESAKTIVITET: 'YRKESAKTIVITET',
   UTSENDING: 'UTSENDING',
   YRKESAKTIVITET_ANTALL_LAND: 'YRKESAKTIVITET_ANTALL_LAND',
-  VIRKSOMHET: 'VIRKSOMHET',
+  VIRKSOMHET_TYPE: 'VIRKSOMHET_TYPE',
   VESENTLIG_VIRKSOMHET: 'VESENTLIG_VIRKSOMHET',
   NORMALT_DRIVER_VIRKSOMHET: 'NORMALT_DRIVER_VIRKSOMHET',
   AKTIVITET: 'AKTIVITET',
   BOSTEDSLAND: 'BOSTEDSLAND',
   TJENESTEMANN: 'TJENESTEMANN',
   FORRETNINGSSTED: 'FORRETNINGSSTED',
-  ARBEIDSGIVERE: 'ARBEIDSGIVERE',
+  VIRKSOMHETER: 'VIRKSOMHETER',
   SOKKEL_SKIP: 'SOKKEL_SKIP',
   VEDTAK: 'VEDTAK',
   ENDRET_PERIODE: 'ENDRE_PERIODE',
